+++ source/include/CommonAssemblyInfo.cs
@@ -2,4 +2,4 @@
 
 [assembly: AssemblyCompany("Codartis")]
 [assembly: AssemblyProduct("NsDepCop")]
-[assembly: AssemblyCopyright("Copyright © 2013-2024 Ferenc Vizkeleti")]
+[assembly: AssemblyCopyright("Copyright © 2013-2025 Ferenc Vizkeleti")]