+++ source/NsDepCop.Test/Interface/Config/DomainTests.cs
@@ -25,7 +25,7 @@ namespace Codartis.NsDepCop.Test.Interface.Config
         }
 
         [Fact]
-        public void Create_WithNull_ThrowsArgumentNullExceptionn()
+        public void Create_WithNull_ThrowsArgumentNullException()
         {
             Assert.Throws<ArgumentNullException>(() => new Domain(null));
         }